pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "nonebot-plugin-endfield"
version = "0.2.0"
authors = [
  { name="NekoTC", email="1719252259@qq.com" },
]
description = "A nonebot plugin for Arknights:Endfield"
readme = "README.md"
requires-python = ">=3.9"
license = "MIT"
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
]
dependencies = [
    "nonebot2>=2.3.0,<3.0.0",
    "nonebot-adapter-onebot>=2.2.2,<3.0.0",
    "nonebot-plugin-apscheduler>=0.3.0,<1.0.0",
    "nonebot-plugin-localstore>=0.7.4,<1.0.0",
    "aiofiles>=23.2.1,<25.0.0",
    "httpx>=0.23.1,<1.0.0",
    "pyecharts>=2.0.0,<3.0.0",
    "playwright>=1.52.0",
    "httpx-ws>=0.7.2,<1.0.0",

    "numpy==1.26.0; python_version < '3.10'",
    "numpy>=2.0.0,<3.0.0; python_version >= '3.10'"
]

[project.urls]
"Homepage" = "https://github.com/TakesBot/nonebot-plugin-endfield"
"Bug Tracker" = "https://github.com/TakesBot/nonebot-plugin-endfield/issues"

[project.entry-points."nonebot.plugin"]
endfield = "endfield"
nonebot_plugin_endfield = "nonebot_plugin_endfield"

[tool.setuptools.packages.find]
include = ["nonebot_plugin_endfield*", "endfield*"]

[tool.setuptools.package-data]
nonebot_plugin_endfield = ["assets/fonts/*"]